Refactor  saleor/graphql/attribute/utils
